feat: consolidate types
