+++ i18n/locales/en.json
@@ -1,4 +1,6 @@
 {
+  "accessDenied": "Access Denied",
+  "accessDeniedMessage": "You are not authorized to view the page you were trying to access. Please contact a Guardian Connector administrator to change your role in order to view this page.",
   "accessPermissions": "Access Permissions",
   "addNewTable": "Add new table",
   "administration": "Administration",
