+++ i18n/locales/nl.json
@@ -99,6 +99,8 @@
   "resetDashboard": "Dashboard resetten",
   "restrictView": "Wilt u deze weergave beperken?",
   "restrictViewDescription": "Wanneer beperkt, kunnen alleen Leden en Beheerders toegang krijgen tot deze weergave. Kijkers worden doorgestuurd naar de hoofdpagina.",
+  "accessDenied": "Toegang Geweigerd",
+  "accessDeniedMessage": "U bent niet gemachtigd om de pagina te bekijken waartoe u probeerde toegang te krijgen. Neem contact op met een Guardian Connector-beheerder om uw rol te wijzigen om deze pagina te kunnen bekijken.",
   "accessPermissions": "Toegangsrechten",
   "satelliteUsedForDetection": "Satelliet gebruikt voor detectie",
   "selectAlertDateRange": "Selecteer een alert datum periode",
