+++ i18n/locales/pt.json
@@ -96,6 +96,9 @@
   "removeTable": "Remover tabela",
   "removeTableAreYouSure": "Tem certeza de que deseja remover esta tabela",
   "resetDashboard": "Redefinir painel",
+  "accessDenied": "Acesso Negado",
+  "accessDeniedMessage": "Você não está autorizado a visualizar a página que estava tentando acessar. Entre em contato com um administrador do Guardian Connector para alterar sua função para visualizar esta página.",
+  "accessPermissions": "Permissões de Acesso",
   "satelliteUsedForDetection": "Satélite usado para detecção",
   "selectAlertDateRange": "Selecione um intervalo de datas de alerta",
   "selectBasemap": "Selecione o mapa base",
